backend/requirements.txt
fastapi
uvicorn
python-dotenv
thisispamela>=1.0.3
pydantic
